requirements.txt
--index-url https://mirrors.aliyun.com/pypi/simple/

pyside6-essentials >= 6.2.0 ; python_version >= "3.9" and python_version < "3.13"
shiboken6 >= 6.2.0 ; python_version >= "3.9" and python_version < "3.13"
pyserial==3.5
pyusb==1.3.1